LayoutTests/media/track/video/video-track-mkv-theora-selected-expected.txt


Test audio track enabled attributes.

EVENT(canplaythrough)
EXPECTED (video.videoTracks.length == '2') OK
RUN(track1 = video.videoTracks[0])
RUN(track2 = video.videoTracks[1])

** Checking initial state
EXPECTED (track1.selected == 'true') OK
EXPECTED (track2.selected == 'false') OK
EXPECTED (video.videoTracks.selectedIndex == '0') OK

** If we select one track, the other should be deselected
RUN(track2.selected = true)
EXPECTED (track1.selected == 'false') OK
EXPECTED (track2.selected == 'true') OK
EXPECTED (video.videoTracks.selectedIndex == '1') OK
END OF TEST
